Doing transfers

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/JdbcAccountDao.java
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
+import java.math.BigDecimal;
 
 import javax.sql.DataSource;
 import java.sql.ResultSet;
@@ -21,36 +22,15 @@ public class JdbcAccountDao implements AccountDao {
     }
 
     @Override
-//    public double getBalance(int userId) throws SQLException {
-//        double balance = null;
-//        String sql = "SELECT balance FROM account WHERE user_id = ?";
-//      SqlRowSet result = jdbcTemplate.queryForRowSet(sql, userId);
-////        balance = jdbcTemplate.queryForObject(sql, double.class, userId);
-////        if (balance.equals(null)) {
-////            return double.valueOf(Long.parseLong("1000"));
-////        }
-//        while (result.next()) {
-//            balance = result.getdouble("balance");
-//        }
-//        balance = getdoubleSafe(result, "balance");
-//        return balance;
-//    }
-
     public double getBalance(int userId) throws SQLException {
         double balance = 0;
         String sql = "SELECT balance FROM account WHERE user_id = ?";
         SqlRowSet result = jdbcTemplate.queryForRowSet(sql, userId);
-//        balance = jdbcTemplate.queryForObject(sql, double.class, userId);
-////        if (balance.equals(null)) {
-////            return double.valueOf(Long.parseLong("1000"));
-//        }
         while (result.next()) {
             balance = result.getDouble("balance");
         }
-//        balance = getdoubleSafe(result, "balance");
         return balance;
     }
-
 
     @Override
     public Account getAccountByUserId(int userId) {
@@ -73,7 +53,7 @@ public class JdbcAccountDao implements AccountDao {
     private Account mapRowToAccount(SqlRowSet rowSet) {
         Account account = new Account();
         account.setAccountId(rowSet.getInt("account_id"));
-        account.setBalance(rowSet.getDouble("balance"));
+        account.setBalance(rowSet.getBigDecimal("balance"));
         return account;
     }
 }
